move drainer batch test into separate file
